requirements.txt
asgiref==3.3.1
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
Pillow==8.1.0
PyJWT==2.0.0
pytz==2020.5
sqlparse==0.4.1
